requirements.txt
flask~=2.2.3
sqlalchemy
flask-sqlalchemy
pymysql
cryptography==40.0.2
Werkzeug~=2.2.3
WTForms~=3.0.1
mysql-connector-python==8.0.33